+++ src/main/kotlin/quickfix/config/RedisConfig.kt
@@ -9,7 +9,7 @@ import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer
 import org.springframework.data.redis.serializer.StringRedisSerializer
 import quickfix.dto.job.jobOffer.CreateJobOfferDTO
 import quickfix.dto.job.jobRequest.JobRequestDTO
-import quickfix.dto.message.RedisMessageDTO
+import quickfix.dto.chat.RedisMessageDTO
 
 @Configuration
 class RedisConfig {
